repl: refactor subscription to interface
